+++ src/main/java/com/stockexchange/matchingservice/model/dto/OrderResponse.java
@@ -3,7 +3,7 @@ package com.stockexchange.matchingservice.model.dto;
 import com.stockexchange.matchingservice.model.Order;
 import com.stockexchange.matchingservice.model.OrderStatus;
 import com.stockexchange.matchingservice.model.OrderType;
-
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
 
@@ -12,7 +12,7 @@ public record OrderResponse(
         OrderStatus orderStatus,
         OrderType orderType,
         String symbol,
-        double price,
+        BigDecimal price,
         int executedQuantity,
         int totalQuantity,
         Instant orderDate,
